rebase with authenticate middleware
