+++ pkg/quarterdeck/middleware/auth.go
@@ -15,7 +15,7 @@ import (
 const (
 	bearer                    = "Bearer "
 	authorization             = "Authorization"
-	UserClaims                = "user_claims"
+	ContextUserClaims         = "user_claims"
 	DefaultKeysURL            = "https://auth.ensign.app/.well-known/jwks.json"
 	DefaultAudience           = "https://ensign.app"
 	DefaultIssuer             = "https://auth.ensign.app"
@@ -71,9 +71,31 @@ func Authenticate(opts ...AuthOption) (_ gin.HandlerFunc, err error) {
 		}
 
 		// Add claims to context for use in downstream processing and continue handlers
-		c.Set(UserClaims, claims)
+		c.Set(ContextUserClaims, claims)
 		c.Next()
 	}, nil
+}
+
+// Authorize is a middleware that requires specific permissions in an authenticated
+// user's claims. If those permissions do not match or the request is unauthenticated
+// the middleware returns a 401 response. The Authorize middleware must be chained
+// following the Authenticate middleware.
+func Authorize(permissions ...string) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		claims, err := GetClaims(c)
+		if err != nil {
+			c.Error(err)
+			c.AbortWithStatusJSON(http.StatusUnauthorized, api.ErrorResponse(ErrNoAuthorization))
+			return
+		}
+
+		if !claims.HasAllPermissions(permissions...) {
+			c.AbortWithStatusJSON(http.StatusUnauthorized, api.ErrorResponse(ErrNoPermission))
+			return
+		}
+
+		c.Next()
+	}
 }
 
 // GetAccessToken retrieves the bearer token from the authorization header and parses it
@@ -93,6 +115,17 @@ func GetAccessToken(c *gin.Context) (tks string, err error) {
 		return "", ErrParseBearer
 	}
 	return "", ErrNoAuthorization
+}
+
+// GetClaims fetches and parses Quarterdeck claims from the gin context. Returns an
+// error if no claims exist on the context; panics if the claims are not the correct
+// type -- however the panic should be recovered by middleware.
+func GetClaims(c *gin.Context) (*tokens.Claims, error) {
+	claims, exists := c.Get(ContextUserClaims)
+	if !exists {
+		return nil, ErrNoClaims
+	}
+	return claims.(*tokens.Claims), nil
 }
 
 // AuthOption allows users to optionally supply configuration to the Authorization middleware.
